Fuking listView

+++ app/src/main/java/pt/unl/fct/data/model/Restaurant.java
@@ -3,9 +3,9 @@ package pt.unl.fct.data.model;
 public class Restaurant {
     int restaurantImage;
     String restaurantName, restaurantRating, restaurantPrice;
-    float latitude, longitude;
+    double latitude, longitude;
 
-    public Restaurant(int restaurantImage, String restaurantName, String restaurantRating, String restaurantPrice, float latitude, float longitude) {
+    public Restaurant(int restaurantImage, String restaurantName, String restaurantRating, String restaurantPrice, double latitude, double longitude) {
         this.restaurantImage = restaurantImage;
         this.restaurantName = restaurantName;
         this.restaurantRating = restaurantRating;
